ditto/requirements.txt
flask==1.1.1
slackclient==2.4.0